requirements.txt
streamlit==1.28.1
pillow==10.0.1
numpy==1.24.3
plotly==5.17.0